Bring in the shared types and assign ICommand
