+++ src/ServiceControl/HeartbeatMonitoring/InternalMessages/RegisterEndpointStartup.cs
@@ -4,17 +4,6 @@
     using System.Collections.Generic;
     using NServiceBus;
 
-    class EndpointHeartbeat : ICommand
-    {
-        public DateTime ExecutedAt { get; set; }
-
-        public string EndpointName { get; set; }
-
-        public Guid HostId { get; set; }
-
-        public string Host { get; set; }
-    }
-
     class RegisterEndpointStartup : ICommand
     {
         public Guid HostId { get; set; }
