+++ src/ServiceControl/SagaAudit/Messages/SagaChangeInitiator.cs
@@ -1,8 +1,9 @@
 namespace ServiceControl.EndpointPlugin.Messages.SagaState
 {
     using System;
+    using NServiceBus;
 
-    class SagaChangeInitiator
+    class SagaChangeInitiator: ICommand
     {
         public string InitiatingMessageId { get; set; }
         public string MessageType { get; set; }
